requirements.txt
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.7.0
aiohttp==3.7.4
aiormq==3.3.1
alembic==1.4.3
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
baidu-aip==2.2.18.0
bidict==0.21.2
boto3==1.18.32
botocore==1.21.32
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorclass==2.2.0
coloredlogs==15.0.1
colorhash==1.0.3
comtypes==1.1.10
constantly==15.1.0
croniter==1.0.15
cryptography==3.4.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.6
dnspython==1.16.0
docker==5.0.0
docopt==0.6.2
english==2020.7.0
fbmessenger==6.0.0
Flask==2.0.1
Flask-SocketIO==4.3.1
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.20
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.3.0
httpx==0.15.4
humanfriendly==9.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.8.1
incremental==21.3.0
isodate==0.6.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mattermostwrapper==2.2
monotonic==1.6
multidict==4.7.6
neotime==1.7.4
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pansi==2020.7.3
parsel==1.6.0
pika==1.2.0
Pillow==8.3.1
priority==1.3.0
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.17.3
psycopg2-binary==2.8.6
py2neo==2020.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///D:/PycharmProjects/Medical_Robot_Rasa/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser==2.20
PyDispatcher==2.0.5
pydot==1.4.2
Pygments==2.10.0
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
pyreadline==2.1
pyrsistent==0.18.0
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==4.6.0
pyttsx3==2.90
pytz==2021.1
pywin32==227
PyYAML==5.4.1
questionary==1.5.2
queuelib==1.6.2
rasa==2.6.2
rasa-sdk==2.8.1
rasa-x @ file:///D:/PycharmProjects/Medical_Robot_Rasa/rasa_x-0.40.1-py3-none-any.whl
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.15.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.7.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.7.1
Scrapy==2.5.0
sentry-sdk==0.19.5
service-identity==21.1.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.4
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.2.0
tqdm==4.59.0
twilio==6.50.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
typeguard==2.12.1
typing-extensions==3.10.0.1
tzlocal==2.1
ujson==1.35
urllib3==1.26.6
w3lib==1.22.0
wcwidth==0.2.5
webexteamssdk==1.6
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
zope.interface==5.4.0
